add utils.js file to put inside it any general utility function
